Added documentation, minor refactoring.
